Fix layout warnings caused by deprecated syntax.
