+++ RvChecklist/Model/ChecklistItem.swift
@@ -8,11 +8,11 @@
 import SwiftUI
 
 struct ChecklistItem: Identifiable, Hashable, Codable {
-    let id: String
-    let name: String
-    let category: String
-    let order: Int
-    let description: String
+    let id: String          // Used by device
+    let name: String        // Title
+    let category: String    // Pre-Trip, Departure, Arrival
+    let order: Int          // Display sort order
+    let description: String // Markdown?
     var isDone: Bool
     
     private var imageName: String
